requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
coverage==7.13.0
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime>=1.17.0
numpy>=1.26.0
pillow>=10.0.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.0
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
urllib3==2.6.2
